+++ packages/sdk/src/index.ts
@@ -65,9 +65,24 @@ export interface RaribleOrderSdk {
 	/**
 	 * Fill order (buy or accept bid - depending on the order type)
 	 *
+	 * @deprecated Use {@link buy} or {@link acceptBid} instead
 	 * @param request order and parameters (amount to fill, fees etc)
 	 */
 	fill: FillOrderAction
+
+	/**
+	 * Buy order
+	 *
+	 * @param request order and parameters (amount to fill, fees etc)
+	 */
+	buy: FillOrderAction
+
+	/**
+	 * Accept bid order
+	 *
+	 * @param request order and parameters (amount to fill, fees etc)
+	 */
+	acceptBid: FillOrderAction
 
 	/**
    * Get fill transaction data (for external sending)
@@ -176,6 +191,8 @@ export function createRaribleSdk(
 			sell: sellService.sell,
 			sellUpdate: sellService.update,
 			fill: filler.fill,
+			buy: filler.buy,
+			acceptBid: filler.acceptBid,
 			getFillTxData: filler.getTransactionData,
 			bid: bidService.bid,
 			bidUpdate: bidService.update,