+++ packages/sdk/src/nft/sign-nft.test.ts
@@ -1,7 +1,7 @@
-import { createGanacheProvider } from "@rarible/ethereum-sdk-test-common"
 import Web3 from "web3"
 import { Web3Ethereum } from "@rarible/web3-ethereum"
 import { toAddress, toBigNumber } from "@rarible/types"
+import { createGanacheProvider } from "@rarible/ethereum-sdk-test-common/build/create-ganache-provider"
 import type { SimpleLazyNft } from "./sign-nft"
 import { signNft } from "./sign-nft"