+++ packages/sdk/src/order/fill-order/rarible-v1.test.ts
@@ -75,7 +75,7 @@ describe("test exchange v1 order", () => {
 			.send({from: seller })
 
 		const signedOrder: SimpleLegacyOrder = { ...order, signature: await sign(order) }
-		await filler.fill({ order: signedOrder, amount: 1, originFee: 100 })
+		await filler.buy({ order: signedOrder, amount: 1, originFee: 100 })
 
 		const ownership = await retry(10, 4000, async () => {
 			const ownership = await ownershipApi.getNftOwnershipById({