+++ packages/sdk/src/order/fill-order/rarible-v2.test.ts
@@ -1,7 +1,8 @@
 import { randomAddress, randomWord, toAddress, toBigNumber } from "@rarible/types"
 import { Web3Ethereum } from "@rarible/web3-ethereum"
 import Web3 from "web3"
-import { awaitAll, createGanacheProvider } from "@rarible/ethereum-sdk-test-common"
+import { awaitAll } from "@rarible/ethereum-sdk-test-common"
+import { createGanacheProvider} from "@rarible/ethereum-sdk-test-common/build/create-ganache-provider"
 import { toBn } from "@rarible/utils/build/bn"
 import { Configuration, GatewayControllerApi } from "@rarible/ethereum-api-client"
 import { send as sendTemplate, sentTx, simpleSend } from "../../common/send-transaction"
@@ -24,7 +25,7 @@ import { getApiConfig } from "../../config/api-config"
 import { RaribleV2OrderHandler } from "./rarible-v2"
 import { OrderFiller } from "./index"
 
-describe("fillOrder", () => {
+describe("buy & acceptBid orders", () => {
 	const { addresses, provider } = createGanacheProvider()
 	const [sender1Address, sender2Address] = addresses
 	const web3 = new Web3(provider as any)
@@ -144,7 +145,7 @@ describe("fillOrder", () => {
 		const signature = await signOrder(ethereum2, config, left)
 
 		const finalOrder = { ...left, signature }
-		await filler.fill({ order: finalOrder, amount: 2, payouts: [], originFees: [] })
+		await filler.buy({ order: finalOrder, amount: 2, payouts: [], originFees: [] })
 
 		expect(toBn(await it.testErc20.methods.balanceOf(sender2Address).call()).toString()).toBe("4")
 		expect(toBn(await it.testErc1155.methods.balanceOf(sender1Address, 1).call()).toString()).toBe("2")
@@ -230,7 +231,7 @@ describe("fillOrder", () => {
 			account: randomAddress(),
 			value: 100,
 		}]
-		await filler.fill({ order: finalOrder, amount: 2, originFees })
+		await filler.buy({ order: finalOrder, amount: 2, originFees })
 
 		expect(toBn(await it.testErc1155.methods.balanceOf(sender2Address, 1).call()).toString()).toBe(
 			before2.minus(2).toFixed()
@@ -240,7 +241,7 @@ describe("fillOrder", () => {
 		)
 	})
 
-	test("should fill order with crypto punks asset", async () => {
+	test("should fill order (buy) with crypto punks asset", async () => {
 		const punkId = 43
 		//Mint punks
 		await sentTx(it.punksMarket.methods.getPunk(punkId), {from: sender2Address})
@@ -283,14 +284,14 @@ describe("fillOrder", () => {
 
 
 		const finalOrder = { ...left, signature }
-		await filler.fill({ order: finalOrder, amount: 1, originFees: []})
+		await filler.buy({ order: finalOrder, amount: 1, originFees: []})
 
 		const ownerAddress = await it.punksMarket.methods.punkIndexToAddress(punkId).call()
 
 		expect(ownerAddress.toLowerCase()).toBe(sender1Address.toLowerCase())
 	})
 
-	test("should fill bid with crypto punks asset", async () => {
+	test("should accept bid with crypto punks asset", async () => {
 		const punkId = 50
 		//Mint crypto punks
 		await sentTx(it.punksMarket.methods.getPunk(punkId), {from: sender2Address})
@@ -339,7 +340,7 @@ describe("fillOrder", () => {
 		const v2Handler = new RaribleV2OrderHandler(ethereum2, simpleSend, config)
 		const filler = new OrderFiller(ethereum2, null as any, v2Handler, null as any, null as any)
 
-		await filler.fill({ order: finalOrder, amount: 1, originFees: []})
+		await filler.acceptBid({ order: finalOrder, amount: 1, originFees: []})
 
 		const ownerAddress = await it.punksMarket.methods.punkIndexToAddress(punkId).call()